requirements.txt
Flask
Flask-MySQLdb
Flask-Bcrypt
python-dotenv